+++ packages/patternfly-4/react-core/src/components/Select/index.d.ts
@@ -1,2 +1,4 @@
 export { default as Select, SelectProps, SelectVariant } from './Select';
 export { default as SelectOption, SelectOptionProps } from './SelectOption';
+export { default as CheckboxSelectOption, CheckboxSelectOptionProps } from './CheckboxSelectOption';
+export { default as CheckboxSelectGroup, CheckboxSelectGroupProps } from './CheckboxSelectGroup';