+++ packages/patternfly-4/react-core/src/components/Tabs/index.d.ts
@@ -1,2 +1,3 @@
 export { default as Tab, TabProps } from './Tab';
-export { default as Tabs, TabsProps } from './Tabs';
+export { default as Tabs, TabsProps, TabsVariant } from './Tabs';
+export { default as TabContent, TabContentProps } from './TabContent';