+++ packages/patternfly-4/react-core/src/components/Tabs/index.js
@@ -1,2 +1,3 @@
 export { default as Tab } from './Tab';
-export { default as Tabs } from './Tabs';
+export { default as Tabs, TabsVariant } from './Tabs';
+export { default as TabContent } from './TabContent';