+++ packages/patternfly-4/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@1.1.10...@patternfly/react-inline-edit-extension@1.1.11) (2019-04-02)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [1.1.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@1.1.9...@patternfly/react-inline-edit-extension@1.1.10) (2019-04-02)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [1.1.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@1.1.8...@patternfly/react-inline-edit-extension@1.1.9) (2019-04-02)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [1.1.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@1.1.7...@patternfly/react-inline-edit-extension@1.1.8) (2019-04-02)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [1.1.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@1.1.6...@patternfly/react-inline-edit-extension@1.1.7) (2019-04-01)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension